Update secret schema work
